alignment of RNA seq data using STAR
I don't think this is a shell script. I think I need to know how to set variables or inputs (in this case input files, ending with .fastq.gz and prefix, depicted as Q_NTG_129.) for shell script. If you can show me online tutorial for this job or tell me how to do this, I'll really appreciate.